Renaming and adding generic functions
